SPDM Get Capability Implementation
Implemented getCapability which returns capability flags
to the application. The capability flag is used while
decoding the measuremement response.

Tested
case 1) getCapabilities after getMeasurements
After, getMeasurements when getCaps is called the
required result is obtained as Version, Caps and Algo(VCA)
is executed

case 2) getCapability when VCA is NOT executed
As expected when VCA is NOT executed 0 is returned

Both the cases are tested with BMC as requester and
Aardvark as responder. Test passed

Signed-off-by: Prithvi A Pai <prithvi.a.pai@intel.com>