Fix for Register Responder Issue (#75)
BMC acts as Responder and PFR as Requester. Hence, SPDM Application
needs to notify MCTP message type SPDM to wrapper. This needs to happen
only in Responder case and NOT in Requester case. Hence introduced a
flag to Resgister only in Responder case

Tested:
BMC as Requester/Responder and Aardvark as Requester/Responder and test
passes

---------

Signed-off-by: Prithvi A Pai <prithvi.a.pai@intel.com>