spdmapplib: Fix timing issue (#89)
Following changes are performed as part of timing issues seen in SPDM
when tested with PFR.
1) Certificate chain was previously read from file system each time.
This caused timing issue. Added fix to read cert chain at the init of
the spdm.
2) APIs which are not part of POR is defined to return false when
invoked. This is because SPDM 1.0 doesn't require implementation for
those APIs.
3) Measurement is read at the init for spdm and stored for the purpose
of post boot attestation. Earlier measurement was read each time causing
delay in response.
4) Private key is read during of init instead of runtime to avoid file
access as it causes delay in response.

Tested:
1) BMC build works fine
2) Timing issue with PFR is resolved

---------

Signed-off-by: Prithvi A Pai <prithvi.a.pai@intel.com>